feat: Abstract method in query class. Range query point distribution untested for now
